\name{CLX.sim.unequalcov}
\alias{CLX.sim.unequalcov}
\title{Cai, Liu, and Xia equal means test simulator under unequal covariances
}
\description{Performs the test in Cai, Liu, and Xia (2014) for testing equality of two mean vectors on multiple data sets generated by \code{\link{build2popData}} when the population covariances matrices are assumed to be unequal.

}
\usage{
CLX.sim.unequalcov(DATA)
}

\arguments{
  \item{DATA}{an object returned by \code{\link{build2popData}}.
}
}
\value{ A data frame with columns
	\item{TSvalues}{values of the test statistic.
	}
	\item{pvalues}{the p values.
	}
}
\references{ 

	Cai, T. T., Liu, W. & Xia, Y. (2014). Two-sample test of high dimensional means under dependence.
		\emph{J. R. Statist. Soc. B.}
}
\author{
Karl Gregory   \email{kgregory@mail.uni-mannheim.de}, \url{http://www.stat.tamu.edu/~kbgregory}.
}

\seealso{
\code{\link{CLX.test.unequalcov}}
}
\examples{
	
\dontrun{

DATA <-build2popData(	
	n = 15,
	m = 20,
	p = 500,
	muX = rep(0,500),
	muY = rep(0,500),
	commoncov = FALSE,
	VarScaleY = 1,
	dep = "ARMA",
	ARMAparms = list(coefs=list(ma=c(.2,.3) , ar=c(.4,-.1))),
	LRparm = .75,
	S = 25,
	innov = function(n,...) rnorm(n,0,1),
	heteroscedastic=TRUE,
	het.diag = diag(.1 + rexp(500,1/2))
	)
	
CLX.sim.unequalcov(DATA)

}
}

\keyword{htest}
\keyword{multivariate}
